[WIP] vold_decrypt:
* Separate stdout and stderr buffers
vdc's return codes get sent to stdout, but the
possible presence of other error messages in
the output buffer will cause a valid return from
vdc not to be parsed properly, and subsequent
decryption to fail due to "misunderstood" return
code.
eg on the U11+ (htc_ocm) libc will generate an
error to stderr due to a missing property area
resulting a proper connection to vold to be
incorrectly parsed, and break decryption.

* Improve logging

Change-Id: I57987ebe4ee6754a78e79ca177506098f8301f8f